requirements.txt
scikit-learn
python-telegram-bot
aiosqlite==0.21.0
flask
numpy==2.3.2
